refactor to 'Jukeboxer'
