+++ src/main/java/jukeboxer/Constants.java
@@ -1,4 +1,4 @@
-package jukebox;
+package jukeboxer;
 
 import net.minecraft.client.audio.MusicTicker.MusicType;
